% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajouter_donnees_assecs_aux_stations.R
\name{ajouter_donnees_assecs_aux_stations}
\alias{ajouter_donnees_assecs_aux_stations}
\title{Ajout des données d'assecs aux stations}
\usage{
ajouter_donnees_assecs_aux_stations(stations_geo, assecs_df)
}
\arguments{
\item{stations_geo}{L'objet géographique points des stations}

\item{assecs_df}{Le dataframe contenant les données d'assecs}
}
\value{
La fonction renvoit l'objet géographique points des stations auquel ont été ajoutées les données
    d'assecs.
}
\description{
Cette fonction permet d'ajouter à l'objet géographique (classe `sf` points) des stations les attributs
    caractérisant les assecs. #'     La correspondance se fait par une jointure sur les codes des
    stations CdSiteHydro.
}
\examples{
\dontrun{
stations_onde_geo <- stations_onde_geo \%>\%
    ajouter_donnees_assecs_aux_stations(assecs_df = assecs)
}
}
