% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajouter_zones_propluvia.R
\name{ajouter_zones_propluvia}
\alias{ajouter_zones_propluvia}
\title{Ajouter les données de zones d'alerte Propluvia au tableau de données onde}
\usage{
ajouter_zones_propluvia(onde_df, propluvia_shape)
}
\arguments{
\item{onde_df}{un tableau de données onde téléchargé avec la
fonction \code{telecharger_donnees_onde_api}}

\item{propluvia_shape}{la couche spatiale pour les zones d'alerte Propluvia.
Une version existe en tant que jeux de données dans le package : (\code{propluvia_zone}).}
}
\value{
un dataframe comprenant les données onde et les données Propluvia jointées spatialement.
}
\description{
Ajouter les données de zones d'alerte Propluvia au tableau de données onde
}
\examples{
\dontrun{
ajouter_zones_propluvia(onde_df = onde_14, propluvia_shape = propluvia_zone)
}
}
