% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembler_fichiers_onde_annuels_csv.R
\name{assembler_fichiers_onde_annuels_csv}
\alias{assembler_fichiers_onde_annuels_csv}
\title{Assemblage des fichiers annuels}
\usage{
assembler_fichiers_onde_annuels_csv(annual_onde_files_dir)
}
\arguments{
\item{annual_onde_files_dir}{Chaîne de caractères indiquant le chemin vers le répertoire où se trouvent
les fichiers annuels}
}
\value{
Dataframe national pour l'ensemble des années de données disponibles
}
\description{
Cette fonction permet d'assembler - en les empilant - les fichiers csv issus de la fonction
    telecharger_fichiers_onde_annuels() en un unique dataframe. La sortie est le dataframe national
    pour l'ensemble des années où la donnée est disponible.
}
\examples{

}
