% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculer_assecs_ete.R
\name{calculer_assecs_ete}
\alias{calculer_assecs_ete}
\title{Caractérisation des assecs}
\usage{
calculer_assecs_ete(onde_df)
}
\arguments{
\item{onde_df}{Le dataframe contenant les données ONDE en format long.}
}
\value{
Un dataframe donnant pour chaque station (ligne) le nombre d'observations, le nombre d'assecs
    et le pourcentage des observations où la station était en assec.
}
\description{
Caractérisation de la fréquence des assecs. Comme le recours à des campagnes dites "complémentaires" varie largement
    selon les régions et départements, pour que la caractérisation ait un sens, elle est restreinte
    aux observations des mois d'été, donc des données des campagnes "usuelles".
}
\examples{
\dontrun{
assecs <- calculer_assecs_ete(onde_df = onde)
}

}
