% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculer_assecs_heatmap.R
\name{calculer_assecs_heatmap}
\alias{calculer_assecs_heatmap}
\title{Fonction pour calculer les pourcentages de stations en assec}
\usage{
calculer_assecs_heatmap(onde_df)
}
\arguments{
\item{onde_df}{un tableau de données onde téléchargé avec la
fonction \code{telecharger_donnees_onde_api}}
}
\value{
un tableau résumant au format pour le graphique avec la
 fonction \code{produire_graph_heatmap_assecs}
}
\description{
Calcule les pourcentages de stations en assec (sur les mois de campagnes usuelles)
 et met en forme du tableau pour le graphique de type \code{heatmap}
}
\examples{
\dontrun{
onde_14 <- telecharger_donnees_onde_api(dpt = c('14'))
calculer_assecs_heatmap(onde_df = onde_14)
}
}
