% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculer_bilan_ecoulement.R
\name{calculer_bilan_ecoulement}
\alias{calculer_bilan_ecoulement}
\title{Calculer un bilan des types d'écoulement pour les campagnes ONDE usuelles selon
 la typologie nationale ou départementale}
\usage{
calculer_bilan_ecoulement(
  onde_df,
  referentiel_onde = "Typologie nationale",
  mod,
  mod_levels,
  force_complementaire = F,
  ...
)
}
\arguments{
\item{onde_df}{un tableau de données onde téléchargé avec la fonction \code{telecharger_donnees_onde_api}}

\item{referentiel_onde}{le nom de la typologie à utiliser,
soit \code{Typologie nationale} ou \code{Typologie départementale}}

\item{mod}{Le nom pour la colonne créée pour la typologie. Par exemple, soit
\code{lib_ecoul3mod} pour la typologie nationale ou \code{lib_ecoul4mod} pour la typologie départementale,}

\item{mod_levels}{les niveaux de facteurs souhaités en fonction des typologies.}

\item{force_complementaire}{booléen, si TRUE les complémentaires sont prises en compte dans le calcul. Par défaut, FALSE.}

\item{...}{paramètre optionnel}
}
\value{
un dataframe avec les pourcentages par types d’écoulement selon la typologie choisie
}
\description{
Calculer un bilan des types d'écoulement pour les campagnes ONDE usuelles selon
 la typologie nationale ou départementale
}
\examples{
\dontrun{
## bilan écoulement sur typologie nationale sur les usuelles uniquement
calculer_bilan_ecoulement(onde_df = onde_df,
                        mod = lib_ecoul,
                        mod_levels = c("Ecoulement visible",
                                       "Ecoulement non visible",
                                       "Assec",
                                       "Observation impossible",
                                       "Donnée manquante"),
                        referentiel_onde = "Typologie nationale",
                        force_complementaire = F)

## bilan écoulement sur typologie départementale avec les complémentaires
calculer_bilan_ecoulement(onde_df = onde_df,
                          mod = lib_ecoul,
                          mod_levels = c("Ecoulement visible acceptable",
                                         "Ecoulement visible faible",
                                         "Ecoulement non visible",
                                         "Assec",
                                         "Observation impossible",
                                         "Donnée manquante"),
                          referentiel_onde = "Typologie départementale",
                          force_complementaire = F)

## bilan écoulement sur typologie nationale avec les complémentaires
calculer_bilan_ecoulement(onde_df = onde_df,
                        mod = lib_ecoul,
                        mod_levels = c("Ecoulement visible",
                                       "Ecoulement non visible",
                                       "Assec",
                                       "Observation impossible",
                                       "Donnée manquante"),
                        referentiel_onde = "Typologie nationale",
                        force_complementaire = T)

}
}
