% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculer_indice_onde.R
\name{calculer_indice_onde}
\alias{calculer_indice_onde}
\title{Fonction pour calculer les indices ONDE sur les campagnes usuelles}
\usage{
calculer_indice_onde(onde_df = NULL, force_complementaire = F)
}
\arguments{
\item{onde_df}{un tableau de données onde téléchargé avec la fonction \code{telecharger_donnees_onde_api}}

\item{force_complementaire}{argument booléen. Si \code{TRUE}, force le calcul des indices sur toutes les campagnes
(usuelles et complémentaires avec un minimum de 25 observations par campagne). Par défaut, égal à \code{FALSE}}
}
\value{
retourne un tableau avec une colonne 'indice'
}
\description{
La fonction permet de calculer les indices ONDE départementaux à partir des campagnes
 usuelles d'un tableau. L'indice est une note variant de 1 à 10, en prenant en compte
 les nombres de stations avec écoulements interrompus et écoulements continus.
}
\examples{
\dontrun{
calculer_indice_onde(onde_df = onde)
}
}
