% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculer_recurrence_assecs.R
\name{calculer_recurrence_assecs}
\alias{calculer_recurrence_assecs}
\title{Fonction pour calculer la récurrence des assecs consécutifs sur les stations}
\usage{
calculer_recurrence_assecs(onde_df)
}
\arguments{
\item{onde_df}{un tableau de données onde téléchargé avec la
fonction \code{telecharger_donnees_onde_api}}
}
\value{
un tableau de synthèse avec le nombre de stations (et pourcentage) par durée
 d'assecs (de 1 à 5 mois consécutifs) sur la période de campagne usuelle (mai à septembre)
}
\description{
Fonction pour calculer la récurrence des assecs consécutifs sur les stations
}
\examples{
\dontrun{
onde_14 <- telecharger_donnees_onde_api(dpt = c('14'))
calculer_recurrence_assecs(onde_df = onde_14)
}

}
