% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completer_observations_mois_manquants.R
\name{completer_observations_mois_manquants}
\alias{completer_observations_mois_manquants}
\title{Ajout des mois sans observations}
\usage{
completer_observations_mois_manquants(onde_df, stations)
}
\arguments{
\item{onde_df}{Le dataframe contenant les données ONDE en format "long".}

\item{stations}{Le vecteur (caractère) des identifiants des stations au sens du champ "CdSiteHydro" de la base ONDE.}
}
\value{
Le dataframe contenant les données ONDE en format "long" complété avec les mois sans données pour lesquels
    l'observation est notée NA.
}
\description{
Afin de permettre une représentation graphique de l'ensemble de la période de collecte des données
    d'étiage, cette fonction ajoute à un tableau de données ONDE, qui ne comprend que des données
    d'observations, tous les mois pour lesquels il n'y a pas eu d'observations. Le tableau de sortie
    comprend donc un nombre de lignes qui est le produit de nombre de stations par le nombre
    d'années d'observations par 12 mois : 12 x nb_années x nb_stations
}
\examples{
\dontrun{
onde_ts_mois <- completer_observations_mois_manquants(onde_df = onde,
stations = codes_stations)}


}
