% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coursdeau_carthage_classe1a4}
\alias{coursdeau_carthage_classe1a4}
\title{Cours d'eau BD Carthage 1 à 4}
\format{
un dataframe au  format \code{sf} (multilignes - crs RGF93)
\describe{
  \item{CdEntiteHy}{code de référence du cours d'eau} (ex: 14 pour le Calvados)
  \item{NomEntiteH}{nom du cours d'eau}
  \item{Classe}{rang du cours d'eau de 1 à 4}
  \item{geometry}{la géométrie du cours d'eau}
}
}
\usage{
coursdeau_carthage_classe1a4
}
\description{
Un tableau contenant les informations géographiques (multilignes) pour les
 cours d'eau de rangs 1 à 4 issus de la BD Carthage.
}
\keyword{datasets}
