% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/departement_shape.R
\docType{data}
\name{departement_shape}
\alias{departement_shape}
\title{Contours des départements français}
\format{
Dataframe au format sf avec 101 lignes et 3 variables:
\describe{
  \item{code insee}{numero de departement, par exemple '14' pour le Calvados}
  \item{nom}{nom du departement}
  \item{geometry}{geometrie des departements}
}
}
\source{
Shapefile contours departements sur le portail data.gouv.fr \url{https://www.data.gouv.fr/fr/datasets/}
}
\usage{
departement_shape
}
\description{
Donnees spatiales au format 'sf', systeme de projection RGF93 / Lambert-93
}
\keyword{datasets}
