% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{departements_FR}
\alias{departements_FR}
\title{Contours des départements français}
\format{
un dataframe au 'sf' (multipolygone) crs RGF93
\describe{
  \item{code_insee}{le code insee des départements}
  \item{nom}{le nom du département}
  \item{geometry}{l'information sur la géométrie du département}
}
}
\usage{
departements_FR
}
\description{
Un tableau contenant les informations géographiques (polygones) des départements
 français. Les variables sont :
}
\keyword{datasets}
