% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporter_les_graphiques_png.R
\name{exporter_les_graphiques_png}
\alias{exporter_les_graphiques_png}
\title{Export des graphiques en format png}
\usage{
exporter_les_graphiques_png(stations_geo, liste_graphiques, repertoire)
}
\arguments{
\item{stations_geo}{La couche géographique des stations (classe `sf`), contenant la variable "CdSiteHydro".
du fichier ONDE.}

\item{liste_graphiques}{La liste contenant les graphiques, qui aura été auparavant produite par la
fonction produire_graph_pour_toutes_les_stations().}

\item{repertoire}{Chaîne de caractère indiquant le chemin vers le répertoire de stockage des fichiers png.}
}
\value{
La fonction produit autant de fichiers png qu'il y a de stations.
}
\description{
Cette fonction exporte chacun des graphiques au format png, le nomme selon les codes SAGEs et station +
    les intitulés des SAGEs. Ils sont stockés dans le répertoire fourni par l'utilisateur. Si le répertoire
    n'existe pas, la fonction le crée. Sinon, elle écrase son contenu antérieur. Au final, il y a un fichier
    par station. L'exécution de la fonction prend environ une seconde par station.
}
\examples{

}
