% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerer_codes_region.R
\name{gerer_codes_region}
\alias{gerer_codes_region}
\title{Gestion des codes régions}
\usage{
gerer_codes_region(onde_df)
}
\arguments{
\item{onde_df}{Un dataframe contenant des données ONDE mis en forme par la fonction
assembler_fichiers_onde_annuels_csv().}
}
\value{
La sortie de la fonction est un dataframe contenant des données ONDE où les codes région
    ont été recodés pour être conformes au découpage régional en 2020.
}
\description{
Gestion des codes régions, en particulier des problèmes issus des fusions qui ont changé les intitulés.
La mise à jour des régions sur les fichiers de données a été opérée début 2020 donc cette fonction
}
\examples{
\dontrun{
onde_bretagne_2019 <- gerer_codes_region(onde_bretagne_2019)}

}
