% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerer_les_campagnes.R
\name{gerer_les_campagnes}
\alias{gerer_les_campagnes}
\title{Gestion des campagnes d'observation de terrain}
\usage{
gerer_les_campagnes(onde_df)
}
\arguments{
\item{onde_df}{Le dataframe contenant les données ONDE mis en forme par la fonction
assembler_fichiers_onde_annuels_csv().}
}
\value{
Le dataframe avec une donnée par mois après application de règles concernant les campagnes
    usuelles et complémentaires.
}
\description{
Avec cette fonction, les données sont filtrées pour ne conserver qu'une observation
    par mois. De mai à septembre inclus, c'est l'observation de la campagne
    "usuelle" qui est retenue. D'octobre à avril, c'est la plus sèche des observations
    "complémentaires".
}
\examples{
\dontrun{
onde <- gerer_les_campagnes(onde_df = onde)}

}
