% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lire_couche_sages.R
\name{lire_couche_sages}
\alias{lire_couche_sages}
\title{Lecture de la couche des SAGEs}
\usage{
lire_couche_sages(fichier_shp, scr = 2154)
}
\arguments{
\item{fichier_shp}{Chaîne de caractères indiquant le chemin vers le fichier shapefile des SAGEs. Il peut être
absolu ou relatif.}

\item{scr}{Numérique entier indiquant le Système de Coordonnées de Référence par son code EPSG.
A titre d'exemples le code du Lambert II étendu est 27572 et celui dui WGS84 est 4326.}
}
\value{
La fonction retourne un objet de classe géographique sf comprenant les polygones correspondant
    aux périmètres des SAGEs.
}
\description{
Cette fonction permet d'importer dans R le fichier shapefile polygones des SAGEs.
    Par défaut, le système de coordonnées (SCR) est le Lambert 93 (EPSG:2154).
}
\examples{
\dontrun{
sages <- lire_couche_sages(fichier_shp = couche_sage, scr = 2154)}

}
