% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{onde_4mod}
\alias{onde_4mod}
\title{Couleurs pour modalités ONDE - typologie 4 modalités}
\format{
un vecteur de couleurs
}
\usage{
onde_4mod
}
\description{
Un vecteur pour les 4 modalités d'observations du protocole ONDE.
}
\keyword{datasets}
