% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{onde_5mod}
\alias{onde_5mod}
\title{Couleurs pour modalités ONDE - typologie 5 modalités}
\format{
un vecteur de couleurs
}
\usage{
onde_5mod
}
\description{
Un vecteur pour les 5 modalités d'observations du protocole ONDE.
}
\keyword{datasets}
