% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onde_bretagne_2019_donnees_test.R
\docType{data}
\name{onde_bretagne_2019}
\alias{onde_bretagne_2019}
\title{Jeu de données d'exemple pour le package ondetools}
\format{
Objet de classe \code{"data.frame"}
}
\source{

}
\usage{
data(onde_bretagne_2019)
}
\description{
Ces données sont les données ONDE nationales téléchargées en mars 2020,
restreintes à la Bretagne et expurgées de variables non essentielles pour
les démonstrations
}
\examples{
data(onde_bretagne_2019)
summary(onde_bretagne_2019)
}
\references{
(\href{https://onde.eaufrance.fr/}{Portail de l'observatoire national des étiages})
}
\keyword{datasets}
\keyword{onde}
