% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passer_en_format_large.R
\name{passer_en_format_large}
\alias{passer_en_format_large}
\title{Passage du tableau de données en format large}
\usage{
passer_en_format_large(onde_df_long)
}
\arguments{
\item{onde_df_long}{dataframe en format long issu de la fonction gerer_les_campagnes()}
}
\value{
le dataframe en format large (une colonne par annee_mois)
}
\description{
Cette fonction "pivote" le tableau ONDE produit par la fonction assembler_fichiers_onde_annuels_csv(),
    qui est en format "long" , pour le mettre en format "large", avec une colonne par combinaison
    annee_mois.
}
\examples{
\dontrun{
onde_lb_large <- passer_en_format_large(onde_df_long = onde)
}

}
