% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produire_carte_statique.R
\name{produire_carte_statique}
\alias{produire_carte_statique}
\title{Produire une carte départementale de la situation mensuelle des observations ONDE}
\usage{
produire_carte_statique(
  onde_df_mois = NULL,
  code_departement = NULL,
  referentiel_onde = "Typologie nationale",
  couleurs = onde_4mod,
  dptFR_shp = NULL
)
}
\arguments{
\item{onde_df_mois}{tableau contenant les données ONDE pour le mois souhaité}

\item{code_departement}{le code insee du département, par exemple '14' pour l'Eure}

\item{referentiel_onde}{la typologie à utiliser,
soit \code{Typologie nationale} ou \code{Typologie départementale}}

\item{couleurs}{un vecteur de couleurs correspondant aux référentiels
national ou départemental}

\item{dptFR_shp}{une couche shape pour les départements FR}
}
\value{
un graphique \code{ggplot2}
}
\description{
Produire une carte départementale de la situation mensuelle des observations ONDE
}
\examples{
\dontrun{
dptFR <-
  COGiter::departements_geo \%>\%
  dplyr::left_join(COGiter::departements \%>\%
                     dplyr::select(DEP, REG, NOM_DEP))

produire_carte_statique(code_departement = '27',
                         onde_df_mois = onde_df_mois2,
                         referentiel_onde = 'Typologie départementale',
                         couleurs = onde_5mod,
                         dptFR_shp = dptFR)

produire_carte_statique(code_departement = '27',
                         onde_df_mois = onde_df_mois2,
                         referentiel_onde = 'Typologie nationale',
                         couleurs = onde_4mod,
                         dptFR_shp = dptFR)
}
}
