% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produire_graph_heatmap_assecs.R
\name{produire_graph_heatmap_assecs}
\alias{produire_graph_heatmap_assecs}
\title{Produire un graphique type heatmap pour les assecs des campagnes usuelles onde}
\usage{
produire_graph_heatmap_assecs(heatmap_df, autoscale = F, rotate = F)
}
\arguments{
\item{heatmap_df}{un dataframe produit avec la fonction \code{calculer_assecs_heatmap}}

\item{autoscale}{booléen, si TRUE l'échelle de couleurs est automatique pour les pourcentages.
Par défaut, autoscale = FALSE (échelle de 0 à 100).}

\item{rotate}{booléen, si TRUE les années sont représentées sur l'axe des ordonnées et
les mois sur l'axe des abscisses. Par défaut, FALSE.}
}
\value{
un graphique \code{ggplot2} type heatmap
}
\description{
Produire un graphique type heatmap pour les assecs des campagnes usuelles onde
}
\examples{
\dontrun{
onde_14 <- telecharger_donnees_onde_api(dpt = c('14'))

heatmap_14 <- calculer_assecs_heatmap(onde_df = onde_14)

produire_graph_heatmap_assecs(heatmap_df = heatmap_14, autoscale = T)

}
}
