% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produire_graph_historique_dpt.R
\encoding{UTF-8}
\name{produire_graph_historique_dpt}
\alias{produire_graph_historique_dpt}
\title{Produire un graphique synthétique sur l'historique des observations pour un département}
\usage{
produire_graph_historique_dpt(onde_df, departement)
}
\arguments{
\item{onde_df}{Le dataframe contenant les données ONDE en format long.}

\item{departement}{Le département, correspondant à son numéro (ex: '14' pour le Calvados)}
}
\value{
un graphique à 3 panels avec :
(i) un panel 'heatmap' représentant la proportion mensuelle des stations en assecs sur la chronique d'observations estivales usuelles (mai à septembre),
(ii) un panel correspondant à la moyenne annuelle des stations en assecs pour les variations interannuelles;
(iii) un panel représentant la durée d'assecs pour les stations du département sur la période d'observation estivales usuelles (mai à septembre).
La durée des assecs pour chaque station est calculée selon le nombre de mois consécutifs où des observations avec la modalité assec est observée.
La plus grande séquence d’observation en assec est gardée pour le calcul de la durée (rq. entre observations successives, l’écoulement peut revenir visible).
Ainsi, les durées en assec peuvent aller de 0 mois (c.à.d. des écoulements toujours visibles et consécutivement observés sur la période estivale) à 5 mois (c.à.d. des assecs toujours observés sur la période estivale).
Le pourcentage représenté pour chaque classe de durée est calculé pour les années d’observations.
}
\description{
Produire un graphique synthétique sur l'historique des observations pour un département
}
\examples{
produire_graph_historique_dpt(onde_df, departement =  '76')

}
