% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produire_graph_indice_mensuel.R
\name{produire_graph_indice_mensuel}
\alias{produire_graph_indice_mensuel}
\title{produire graph inidice mensuel}
\usage{
produire_graph_indice_mensuel(
  onde_df,
  code_dpt,
  mois_cible = 5,
  force_complementaire = FALSE
)
}
\arguments{
\item{onde_df}{dataframe produit avec la fonction d'import des donnees}

\item{code_dpt}{code du departement cible}

\item{mois_cible}{code du mois cible}

\item{force_complementaire}{pour prendre en compte ou pas les complementaires}
}
\value{
graphique ggplot
}
\description{
produire graph inidice mensuel
}
\examples{
produire_graph_indice_mensuel(onde_df, code_dpt = '59', mois_cible = 5, force_complementaire = FALSE)

}
