% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produire_graph_pour_toutes_les_stations.R
\name{produire_graph_pour_toutes_les_stations}
\alias{produire_graph_pour_toutes_les_stations}
\title{Produire les graphiques pour toutes les stations}
\usage{
produire_graph_pour_toutes_les_stations(
  stations,
  fonction_graphique,
  onde_df_ts_mois,
  couleurs
)
}
\arguments{
\item{stations}{Le vecteur (caractère) des identifiants des stations au sens du champ "CdSiteHydro" de la base ONDE.}

\item{fonction_graphique}{La fonction graphique à appliquer sur chacune des stations. Ici il s'agit de la fonction
produire_graph_pour_toutes_les_stations().}

\item{onde_df_ts_mois}{Le dataframe des données ONDE qui a été complété pour les mois où il n'y a pas eu d'observations.
Produit par le fonction completer_observations_mois_manquants().}

\item{couleurs}{Un vecteur nommé qui associe à chacune des modalités d'observation (ex : écoulement visible)
une couleur.}
}
\value{
La fonction produit un objet de classe "list" qui contient autant de graphiques qu'il y a de stations.
}
\description{
Cette fonction est une "moulinette" qui permet d'appliquer la fonction de production du graphique non pas à une
    station, mais à un ensemble.
}
\examples{
\dontrun{
graphiques <- produire_graph_pour_toutes_les_stations(stations = codes_stations,
fonction_graphique = produire_graph_pour_une_station,
onde_df = onde_ts_mois,
couleurs = mes_couleurs)
}
}
