% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produire_graph_pour_une_station.R
\name{produire_graph_pour_une_station}
\alias{produire_graph_pour_une_station}
\title{Produire le graphique des assecs pour une station}
\usage{
produire_graph_pour_une_station(code_station, onde_df, couleurs)
}
\arguments{
\item{code_station}{L'identifiant de la station ONDE (Variable CdSiteHydro).}

\item{onde_df}{Le dataframe en format "long" qui contient les données ONDE.}

\item{couleurs}{Un vecteur nommé qui associe à chacune des modalités d'observation (ex : écoulement visible)
une couleur.}
}
\value{
Le graphique obtenu par ggplot.
}
\description{
Cette fonction permet de produire le graphique des assecs pour une station donnée sur l'ensemble des années
    de la série.
}
\examples{

}
