% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produire_graph_pour_une_station_v2.R
\name{produire_graph_pour_une_station_v2}
\alias{produire_graph_pour_une_station_v2}
\title{Produire graphique historique pour station (nouvelle version!)}
\usage{
produire_graph_pour_une_station_v2(
  station_vec,
  onde_df,
  type_mod,
  mod_levels,
  mod_colors
)
}
\arguments{
\item{station_vec}{identifiant de la station à représenter graphiquement (colonne \code{code_station})}

\item{onde_df}{un tableau de données onde téléchargé avec la
fonction \code{telecharger_donnees_onde_api}}

\item{type_mod}{nom de colonne choisie définissant la variable écoulement}

\item{mod_levels}{vecteur définissant les noms de modalités d'écoulement.
Doit correspondre aux valeurs retrouvées dans la colonne \code{type_mod}}

\item{mod_colors}{vecteur de couleur nommé qui associe à chacune des modalités
d'observation une couleur.}
}
\value{
un graphique \code{ggplot2} pour l'historique des observations à la station
 avec les différentes campagnes (usuelles et complémentaires)
}
\description{
Nouvelle version de la fonction pour produire un graphique à la station de l'historique des écoulements.
Celle-ci permet d'utiliser les données téléchargées issues de l'API
et de pouvoir représenter, à la fois, les campagnes usuelles et complémentaires
sur le même graphique. Il est également possible d'utiliser la typologie départementale.
}
\examples{
\dontrun{
onde_14 <- telecharger_donnees_onde_api(dpt = c('14'))

onde_periode <-
  onde_14 \%>\%
  dplyr::mutate(
    Annee = as.numeric(Annee),
    Mois = format(as.Date(date_campagne), "\%m"),
    Mois_campagne = lubridate::ym(paste0(Annee,Mois,sep="-"))
  ) \%>\%
  dplyr::mutate(
    libelle_ecoulement = dplyr::case_when(
      libelle_ecoulement == 'Ecoulement visible' ~ 'Ecoulement visible acceptable',
      TRUE ~ libelle_ecoulement
    )
  )

produire_graph_pour_une_station_v2(station_vec = onde_periode$code_station[1],
                                type_mod = libelle_ecoulement,
                                onde_df = onde_periode,
                                mod_levels = c("Assec",
                                               "Ecoulement\nnon visible",
                                               "Ecoulement\nvisible\nfaible",
                                               "Ecoulement\nvisible\nacceptable",
                                               "Observation\nimpossible",
                                               "Donnée\nmanquante"),
                                mod_colors = onde_5mod

)
}
}
