% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produire_graph_reccurrence_assecs.R
\name{produire_graph_reccurrence_assecs}
\alias{produire_graph_reccurrence_assecs}
\title{Produire un graphique du nombre de stations réparties selon les durées en
 assecs consécutifs par années de suivis.}
\usage{
produire_graph_reccurrence_assecs(df_assecs, round_prec = 1)
}
\arguments{
\item{df_assecs}{un tableau résumé de la durée des assecs produit avec}

\item{round_prec}{nombre de chiffres pour arrondir la valeur du pourcentage calculé}
}
\value{
un graphique \code{ggplot2}
}
\description{
La durée des assecs, en nombre de mois consécutifs, est calculée pour chaque station sur
 la période des campagnes usuelles (de mai à septembre). Les campagnes complémentaires ne sont
 pas utilisées.
}
\examples{
\dontrun{
onde_14 <- telecharger_donnees_onde_api(dpt = c('14'))

onde_14_recurrence <- calculer_recurrence_assecs(onde_df = onde_14)

produire_graph_reccurrence_assecs(df_assecs = onde_14_recurrence)
}
}
