% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produire_graph_type_ecoulement.R
\name{produire_graph_type_ecoulement}
\alias{produire_graph_type_ecoulement}
\title{Produire le graphique bilan des écoulements}
\usage{
produire_graph_type_ecoulement(data_bilan, lib_ecoulement, historique = FALSE)
}
\arguments{
\item{data_bilan}{un dataframe comprenant les différents types d'écoulement observés.
Peut être calculé avec la fonction \code{calculer_bilan_ecoulement}}

\item{lib_ecoulement}{la colonne contenant la variable des types écoulement}

\item{historique}{booléen. Si FALSE, représente le graphique pour la dernière année en cours
(année maximum des données).
Si TRUE, représente le graphique résumant l'historique pour toutes les années.}
}
\value{
un graphique avec \code{ggplot2}
}
\description{
Produire le graphique bilan des écoulements
}
\examples{
\dontrun{
# Exemple bilan typologie départementale campagnes usuelles pour le 27
data_bilan_ecoulement <-
  calculer_bilan_ecoulement(onde_df = onde_df,
                            mod = lib_ecoul,
                            mod_levels = c("Ecoulement visible acceptable",
                                           "Ecoulement visible faible",
                                           "Ecoulement non visible",
                                           "Assec",
                                           "Observation impossible",
                                           "Donnée manquante"),
                            referentiel_onde = "Typologie départementale",
                            force_complementaire = F) \%>\%
  dplyr::filter(code_departement == 27)

## graph bilan de l'année en cours
produire_graph_type_ecoulement(data_bilan = data_bilan_ecoulement,
                               lib_ecoulement = lib_ecoul,
                               historique = F)

## graph bilan de toutes les années
produire_graph_type_ecoulement(data_bilan = data_bilan_ecoulement,
                               lib_ecoulement = lib_ecoul,
                               historique = T)

# Exemple bilan typologie nationale TOUTES campagnes pour le 27
data_bilan_ecoulement <-
  calculer_bilan_ecoulement(onde_df = onde_df,
                            mod = lib_ecoul,
                            mod_levels = c("Ecoulement visible",
                                           "Ecoulement non visible",
                                           "Assec",
                                           "Observation impossible",
                                           "Donnée manquante"),
                            referentiel_onde = "Typologie nationale",
                            force_complementaire = T) \%>\%
  dplyr::filter(code_departement == 27)

produire_graph_type_ecoulement(data_bilan = data_bilan_ecoulement,
                               lib_ecoulement = lib_ecoul,
                               historique = F)
  }
}
