% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produire_rapport_mensuel_dpt.R
\name{produire_rapport_mensuel_dpt}
\alias{produire_rapport_mensuel_dpt}
\title{Fonction pour produire un rapport mensuel départemental ONDE}
\usage{
produire_rapport_mensuel_dpt(
  onde_df,
  code_departement,
  annee_mois,
  region_dr,
  complementaire = FALSE,
  dossier_sortie = getwd()
)
}
\arguments{
\item{onde_df}{tableau contenant les données ONDE}

\item{code_departement}{un seul code (insee) de département (ex: '27') ou
une liste de départements pour plusieurs rapports (ex: \code{c('14', '27', '76')})}

\item{annee_mois}{l'année et le mois du(des) rapport(s)}

\item{region_dr}{le nom de la DR qui sera affichée sur la page de garde du
rapport}

\item{complementaire}{booléen, si TRUE génère un rapport pour une campagne complémentaire.
par défaut FALSE.}

\item{dossier_sortie}{le dossier de sortie où le rapport sera stocké. Par défaut,
le dossier de travail \code{getwd()}}
}
\value{
Renvoie un rapport mensuel au format Word pour les données du réseau ONDE
 pour le département le mois et l'année spécifié
}
\description{
Fonction pour produire un rapport mensuel départemental ONDE
}
\examples{
## Un rapport pour le département 02 pour le mois de mai 2022 ('2022_05')
\dontrun{
produire_rapport_mensuel_dpt(onde_df = donnees_onde,
                             code_departement = '02',
                             annee_mois = "2022_05",
                             region_dr = 'Hauts-de-France',
                             complementaire = F,
                             dossier_sortie = "./OUTPUT")

Rapports pour plusieurs départements (ex: 14, 27, 76) pour des complémentaires
  du mois de juin 2023 ('2023_06')

produire_rapport_mensuel_dpt(onde_df = donnees_onde,
                             code_departement = c('14', '27', '76'),
                             annee_mois = "2022_06",
                             region_dr = 'Normandie',
                             complementaire = T,
                             dossier_sortie = "./OUTPUT")
}
}
