% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telecharger_donnees_onde_api.R
\name{telecharger_donnees_onde_api}
\alias{telecharger_donnees_onde_api}
\title{Télécharger les données ONDE pour un ou plusieurs départements}
\usage{
telecharger_donnees_onde_api(dpt = NULL)
}
\arguments{
\item{dpt}{un vecteur comprenant le code insee du(des) département(s). ex: \code{c('27')}}
}
\value{
un dataframe avec les données campagnes, observations et stations.
}
\description{
Cette fonction permet de télécharger les données ONDE via l'API Ecoulement sur la plateforme
 HubEau (\url{https://hubeau.eaufrance.fr/page/api-ecoulement})
 pour un ou plusieurs départements.
}
\examples{
\dontrun{

## pour un département

telecharger_donnees_onde_api(dpt= c('14'))

## pour plusieurs départements

telecharger_donnees_onde_api(dpt= c('02', '60', '80'))
}
}
\seealso{
\url{https://hubeau.eaufrance.fr/page/api-ecoulement}
}
