% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telecharger_donnees_onde_api_dates.R
\name{telecharger_donnees_onde_api_dates}
\alias{telecharger_donnees_onde_api_dates}
\title{Télécharger les données ONDE pour un ou plusieurs départements et selection
des dates min et max}
\usage{
telecharger_donnees_onde_api_dates(
  dpt = NULL,
  date_min = "2012-01-01",
  date_max = as.character(format(Sys.time(), "\%Y-\%m-\%d"))
)
}
\arguments{
\item{dpt}{un vecteur comprenant le code insee du(des) département(s). ex: \code{c('27')}}

\item{date_min}{chaine de character avec la date minimale à prendre en compte pour les données ex : \code{"2012-01-01"}
par défault la valeur est le "2012-01-01"}

\item{date_max}{chaine de character avec la date minimale à prendre en compte pour les données ex : \code{"2024-09-06"}
par défault la valeur est la date du jour.}
}
\value{
un dataframe avec les données campagnes, observations et stations.
}
\description{
Cette fonction permet de télécharger les données ONDE via l'API Ecoulement sur la plateforme
 HubEau (\url{https://hubeau.eaufrance.fr/page/api-ecoulement})
 pour un ou plusieurs départements.
}
\examples{
\dontrun{

## pour un département

telecharger_donnees_onde_api_dates(dpt= c('14'), date_min = "2025-07-28")

## pour plusieurs départements

telecharger_donnees_onde_api_dates(dpt= c('02', '60', '80'), date_max = "2025-07-28")
}
}
\seealso{
\url{https://hubeau.eaufrance.fr/page/api-ecoulement}
}
