% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produire_barplot_region_par_dpt.R
\name{produire_barplot_region_par_dpt}
\alias{produire_barplot_region_par_dpt}
\title{produire_barplot_region_par_dpt}
\usage{
produire_barplot_region_par_dpt(
  mois_sel,
  annee_sel,
  region_dr,
  referentiel_onde = "Typologie nationale",
  complementaire = FALSE
)
}
\arguments{
\item{mois_sel}{: Chaine de caracteres avec le numero du mois selectionné ex : "05"}

\item{annee_sel}{: Chaine de caracteres avec l'année selectionnée}

\item{region_dr}{: Chaine de caractere avec le nom de la région souhaité (attention à l’orthographe de la région)}

\item{referentiel_onde}{: Chaiende caracteres avec pour savoir s'il faut utiliser la
typologie nationale ou départementale. ('Typologie nationale' ou 'Typologie departementale')}

\item{complementaire}{: Booleen s'il s'agit d'une campagne complémentaire ou pas.}
}
\value{
ggplot graphique
}
\description{
produire_barplot_region_par_dpt
}
\examples{
produire_barplot_region_par_dpt(mois_sel = "08", annee_sel = "2026", region_dr = 'Grand Est')
}
